C-Kermit 4C(053)+1 21 Jun 85
These files come from a Venix 2.0 for the DEC Professional 350/380 computers
directory from some DECUS tape. Or more accurately from the I've not been able
to idnetify which one yet. I deleted the kermit binary that was also
included. These are from the k4cexe.dsk and k4csrc.dsk. Similar images appeared
in at least the VMSLT2002 tape, but lacked the Kermit files (though there was a
binary of Kermit 4E included). https://www.digiater.nl/openvms/decus/vmslt02a/pro
has an archived copy of that.

I don't think the DECnotes nor the KERMIT_INFO files were originally from the 4C
distribution. DECnotes appear to be document a number of changes to the 4C base
sources, so these may not be pure 4C as released by Columnia, but it's the best
we have. The files that may be modified are:
	ckccmai.c, ckcfns.c ckucon.c, ckudia.c, ckusrc.c, ckutio.c, ckuus2.c
	ckuusr.c, ckvcon.c, ckvio.c, ckvkern.com, ckvmak.com, ckvtio.c
KERMIT_INFO has good data on building in the DEC environment, as you might
expect. It's reported that these changes were submitted to Columbia, but I've not
tried to see if they line up with later versions or not.

Note: files that were in prior releases, but not this one, have not been
deleted.