Also C-Kermit, 4D(061) 8 Sep 86
But As of: 12:07pm  Thursday, 19 March 1987

From the RSX87b DECUS tape (just the ck* files). There's a number of differences
with other 4D(061) distributions, and many of the new files here. These files
are likely for non-unix versions of C-KERMIT. ckv is VMS, ckm is mac, cki is
amiga. These were scrubbed from later versions of C-Kermit, and other
distributions are spotty on including them, so use them with caution for other
revisions.