C-Kermit, 4E(070) 29 Jan 88
This appears to be an early release of Kermit 4E. It differns in a number of
small ways from 4E(072), most of which are changes in 072 (like Copyright notice
updates, addition of code from contributors, their names, etc).

However, this comes from the Dan Cross 10th Edition tape that's in TUHS. It
appears to have added #ifdefs for V9 by Paul Glick, 9th edition Unix
support. These are not in the 072 version.  These files is found at
Distributions/Research/Dan_Cross_v10/v10/lbin/kermit in the archive (again,
expanded from files higher in the tree).