Another version of C-Kermit 5A(179) BETA, 8 Feb 92
This one is ported to 386BSD. There's a lot of copyright churn in this, as well
as a number of weird changes. It appears to be newer than the other 5A beta
version that I have, and there's an odd amount of churn. The README file says
there's an RCS tree that documents the changes, but that's not survived.

This appaers to have come from the historical unix tree, by way of Jonathan
Grays 386BSD tree which has the patchkit 0.2.4 applied on top.